removing support of configfile arg in JS
